# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Bloackstairs",
        "description": "test blockstairs api",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/colors": {
            "get": {
                "tags": [
                    "Colors"
                ],
                "summary": "Get list Colors",
                "operationId": "10fc4e94193fcbdb91530214ca745193",
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "rows": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/color": {
            "post": {
                "tags": [
                    "Colors"
                ],
                "summary": "Create new Color",
                "operationId": "67adb7303156c76e75004ff0ded1e94e",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "description": {
                                                "type": "string"
                                            },
                                            "hex": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "description": "Branco",
                                    "hex": "#ffff"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Color registered successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "example": "color id"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "description color"
                                        },
                                        "hex": {
                                            "type": "string",
                                            "example": "#fffff"
                                        },
                                        "rgb": {
                                            "type": "string",
                                            "example": "rgb(255 255 0 0)"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "invalid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "fail"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/color/{id}": {
            "get": {
                "tags": [
                    "Colors"
                ],
                "summary": "Get Detail Color",
                "operationId": "ece0936c7c5f7220b8d3870690b4bb27",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "example": "color id"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "description color"
                                        },
                                        "hex": {
                                            "type": "string",
                                            "example": "#fffff"
                                        },
                                        "rgb": {
                                            "type": "string",
                                            "example": "rgb(255 255 0 0)"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        },
                                        "colors": {
                                            "type": "string",
                                            "example": "colors[]"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Colors"
                ],
                "summary": "Update Color",
                "operationId": "2508ad3529f4b162f4f2c1f826f85925",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "description": {
                                                "type": "string"
                                            },
                                            "hex": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "description": "example description",
                                    "hex": "example #ffff"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Color successfully changed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "example": "color id"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "description color"
                                        },
                                        "hex": {
                                            "type": "string",
                                            "example": "#fffff"
                                        },
                                        "rgb": {
                                            "type": "string",
                                            "example": "rgb(255 255 0 0)"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        },
                                        "users": {
                                            "type": "string",
                                            "example": "users[]"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Colors"
                ],
                "summary": "Delete Color",
                "operationId": "fdb3f120e0e4c40605699f4543b24284",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "Color successfully deleted"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}